+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased - ReleaseDate
 
 * Add a column `gbrs_paired_asm` to ardb
-* Add `n50.sh` to `nwr assembly`
+* Add `n50.sh` and `finish.sh` to `nwr assembly`
 
 ## 0.5.10 - 2023-01-28